pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "knot"
version = "0.1.0"
description = "Local-first CLI for turning rough notes into linked vault notes."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer[all]>=0.12.3",
    "rich>=13.7.1",
    "python-dotenv>=1.0.1",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-google-genai>=1.0.10",
    "langchain-chroma>=0.1.2",
    "langchain-text-splitters>=0.3.0",
    "chromadb>=0.5.5",
]

[project.scripts]
knot = "main:app"

[tool.setuptools]
py-modules = ["api", "main", "core", "prompts", "models", "librarian"]
